refactor(library): extract history/downloads/subscriptions under ≤1000 LOC
Move History, Downloads, and Subscriptions UI into feature.library
subpackages while keeping Screen façades in the library package.
Refresh the feature:library ktlint baseline for the new paths.
Behavior unchanged.
