+++ feature/library/src/main/java/cx/aswin/boxlore/feature/library/downloads/DownloadModels.kt
@@ -1,0 +1,179 @@
+package cx.aswin.boxlore.feature.library.downloads
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.PlayArrow
+import androidx.compose.material3.Checkbox
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import cx.aswin.boxlore.core.data.database.DownloadedEpisodeEntity
+import cx.aswin.boxlore.core.designsystem.theme.expressiveClickable
+import cx.aswin.boxlore.core.model.Episode
+import cx.aswin.boxlore.core.model.Podcast
+
+fun DownloadedEpisodeEntity.toEpisode() = Episode(
+    id = episodeId,
+    title = episodeTitle,
+    description = episodeDescription ?: "",
+    audioUrl = localFilePath,
+    imageUrl = episodeImageUrl,
+    podcastImageUrl = podcastImageUrl,
+    duration = (durationMs / 1000).toInt(),
+    publishedDate = publishedDate
+)
+
+fun DownloadedEpisodeEntity.toPodcast() = Podcast(
+    id = podcastId,
+    title = podcastName,
+    artist = "",
+    imageUrl = podcastImageUrl ?: "",
+    description = "",
+    genre = ""
+)
+
+internal data class PodcastGroup(
+    val podcastId: String,
+    val podcastName: String,
+    val podcastImageUrl: String?,
+    val episodes: List<DownloadedEpisodeEntity>,
+    val totalSizeBytes: Long,
+    val latestDownloadedAt: Long
+)
+
+internal fun formatSize(bytes: Long): String {
+    if (bytes <= 0) return "0 B"
+    val units = arrayOf("B", "KB", "MB", "GB", "TB")
+    val digitGroups = (Math.log10(bytes.toDouble()) / Math.log10(1024.0)).toInt()
+    return String.format(java.util.Locale.US, "%.1f %s", bytes / Math.pow(1024.0, digitGroups.toDouble()), units[digitGroups])
+}
+
+internal fun formatRelativeDate(timestampSeconds: Long): String {
+    if (timestampSeconds == 0L) return ""
+    val now = System.currentTimeMillis() / 1000
+    val diff = now - timestampSeconds
+    return when {
+        diff < 0 -> "Just now"
+        diff < 3600 -> "${diff / 60}m ago"
+        diff < 86400 -> "${diff / 3600}h ago"
+        diff < 604800 -> "${diff / 86400}d ago"
+        diff < 2592000 -> "${diff / 604800}w ago"
+        diff < 31536000 -> "${diff / 2592000}mo ago"
+        else -> "${diff / 31536000}y ago"
+    }
+}
+
+@Composable
+internal fun PodcastListShowCard(
+    group: PodcastGroup,
+    onClick: () -> Unit,
+    onPlayClick: () -> Unit,
+    modifier: Modifier = Modifier,
+    isSelectionMode: Boolean = false,
+    isSelected: Boolean = false,
+    onSelectedChange: (Boolean) -> Unit = {}
+) {
+    val title = group.podcastName
+    val imageUrl = group.podcastImageUrl
+    val downloadCount = group.episodes.size
+    val totalSizeBytes = group.totalSizeBytes
+    val latestDownloadedAt = group.latestDownloadedAt
+    Row(
+        modifier = modifier
+            .fillMaxWidth()
+            .expressiveClickable(
+                onClick = {
+                    if (isSelectionMode) {
+                        onSelectedChange(!isSelected)
+                    } else {
+                        onClick()
+                    }
+                },
+                shape = RoundedCornerShape(16.dp)
+            )
+            .background(MaterialTheme.colorScheme.surfaceContainerLow, shape = RoundedCornerShape(16.dp))
+            .padding(12.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.spacedBy(16.dp)
+    ) {
+        if (isSelectionMode) {
+            Checkbox(
+                checked = isSelected,
+                onCheckedChange = onSelectedChange
+            )
+        }
+
+        AsyncImage(
+            model = imageUrl,
+            contentDescription = title,
+            contentScale = ContentScale.Crop,
+            modifier = Modifier
+                .size(56.dp)
+                .clip(RoundedCornerShape(10.dp))
+                .border(
+                    width = 1.dp,
+                    color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.08f),
+                    shape = RoundedCornerShape(10.dp)
+                )
+        )
+
+        Column(modifier = Modifier.weight(1f)) {
+            Text(
+                text = title,
+                style = MaterialTheme.typography.titleMedium,
+                fontWeight = FontWeight.Bold,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis
+            )
+            Text(
+                text = "$downloadCount ${if (downloadCount == 1) "episode" else "episodes"} • ${formatSize(totalSizeBytes)}",
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.onSurfaceVariant
+            )
+            val latestRelative = formatRelativeDate(latestDownloadedAt / 1000L)
+            if (latestRelative.isNotEmpty()) {
+                Text(
+                    text = "Latest: $latestRelative",
+                    style = MaterialTheme.typography.labelSmall,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.8f)
+                )
+            }
+        }
+
+        if (!isSelectionMode) {
+            IconButton(
+                onClick = onPlayClick,
+                modifier = Modifier
+                    .size(36.dp)
+                    .background(MaterialTheme.colorScheme.primaryContainer, shape = CircleShape)
+            ) {
+                Icon(
+                    imageVector = Icons.Rounded.PlayArrow,
+                    contentDescription = "Play Show Downloads",
+                    tint = MaterialTheme.colorScheme.onPrimaryContainer,
+                    modifier = Modifier.size(20.dp)
+                )
+            }
+        }
+    }
+}
+
